+++ cmake/NAG.cmake
@@ -6,7 +6,7 @@ set (CRAY_POINTER "")
 set (EXTENDED_SOURCE "-132")
 
 set (check_kinds "-kind=unique")
-set (CMAKE_Fortran_FLAGS_DEBUG "-O0 -g -C=all -Wp,-P") # -C=undefined")
-#set (CMAKE_Fortran_FLAGS_DEBUG "-O0 -g -C=all") # -C=undefined")
-set (CMAKE_Fortran_FLAGS_RELEASE "-O3 -g -Wp,-P")
+set (CMAKE_Fortran_FLAGS_DEBUG "-O0 -g -quiet -C=all -Wp,-P") # -C=undefined")
+#set (CMAKE_Fortran_FLAGS_DEBUG "-O0 -g -quiet -C=all") # -C=undefined")
+set (CMAKE_Fortran_FLAGS_RELEASE "-O3 -g -quiet -Wp,-P")