+++ ChangeLog.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added `-quiet` flag for NAG Fortran
+
 ## [1.8.0] - 2024-03-03
 
 ### Added